Created Safe Event Param lookups
Event->checkStrParam(string, string)
Event->checkNumParam(string, float)

+++ EventQueue.cpp
@@ -45,7 +45,7 @@ void EventQueue::updateEventQueue(void)
 		REMARKS:    First in First out
 	*/
 
-   for( int i = 0; i < eventQueue.size(); i++ ) 
+   for( int i = 0; i < (int) eventQueue.size(); i++ ) 
       notifyListeners(&eventQueue[i]);
    
    eventQueue.clear();
